# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cobol-api",
    "version": "1.0.0",
    "description": "Moteur d'amortissement de pret ecrit en COBOL, expose en REST, pour le marche tunisien.\n\nLes montants sont en dinars, exprimes au millime : le dinar se divise en mille. Ils sont rendus comme des nombres JSON a trois decimales exactes, produits par une arithmetique decimale a virgule fixe, et ne transitent jamais par un flottant binaire.\n\nSix invariants sont tenus : la somme des parts de capital egale exactement le capital emprunte, la somme des echeances egale exactement capital plus interets, le solde apres la derniere echeance est exactement nul, chaque ligne verifie echeance = capital + interets, chaque ligne verifie mensualite = echeance + assurance, et le total verse egale la somme des mensualites. La derniere echeance absorbe le residu d'arrondi.\n\nLe TEG est calcule selon le decret n° 2000-462 du 21 fevrier 2000 : le taux de periode est resolu par methode actuarielle sur l'ensemble des versements dus, interets, frais, commissions et assurance compris, puis annualise de facon proportionnelle et exprime avec deux decimales. Sans frais ni assurance, il egale le taux nominal.\n\nLe verdict de taux excessif applique la loi n° 99-64 du 15 juillet 1999 : est excessif tout pret dont le TEG excede de plus du cinquieme le taux effectif moyen publie pour la categorie de concours.",
    "license": {
      "name": "Proprietaire"
    }
  },
  "servers": [
    {
...
    }
  ],
...
  "security": [
    {
      "cleAPI": []
    }
  ],
  "paths": {
...
    "/v1/baremes": {
      "get": {
        "tags": [
          "credit"
        ],
        "summary": "Bareme en vigueur",
        "description": "Taux effectifs moyens en vigueur, une ligne par categorie de concours. Le seuil du taux excessif n'y figure pas : il se deduit du taux effectif moyen par la regle du cinquieme, portee par le moteur de calcul.",
        "operationId": "bareme",
        "responses": {
          "200": {
            "description": "Bareme en vigueur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "count",
                    "baremes"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "success"
                    },
                    "count": {
                      "type": "integer",
                      "example": 8
                    },
                    "baremes": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Bareme"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Cle d'API absente ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            }
          },
          "503": {
            "description": "Bareme indisponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            }
          }
        }
      }
    },
    "/v1/simulations": {
      "get": {
        "tags": [
          "service"
        ],
        "summary": "Piste d'audit",
        "description": "Echeanciers produits, du plus recent au plus ancien. Chaque echeancier produit y laisse une trace : l'ecriture est synchrone et son echec fait echouer la requete, une piste d'audit a trous n'en etant pas une.",
        "operationId": "simulations",
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          },
          {
            "name": "non_conformes",
            "in": "query",
            "required": false,
            "description": "A true, restreint aux prets juges excessifs.",
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Piste d'audit",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "count",
                    "simulations"
                  ],
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "success"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "simulations": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Simulation"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Limite hors bornes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            }
          },
          "401": {
            "description": "Cle d'API absente ou invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            }
          },
          "503": {
            "description": "Piste d'audit indisponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erreur"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/connexion": {
      "post": {
        "summary": "Ouvrir une session d'agent",
        "description": "Rend un jeton porteur valable huit heures. L'identifiant inconnu et le mot de passe faux rendent le meme message : les distinguer permettrait d'enumerer les comptes.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "identifiant",
                  "mot_de_passe"
                ],
                "properties": {
                  "identifiant": {
                    "type": "string"
                  },
                  "mot_de_passe": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session ouverte"
          },
          "401": {
            "description": "Identifiant ou mot de passe incorrect"
          }
        }
      }
    },
    "/v1/auth/deconnexion": {
      "post": {
        "summary": "Revoquer la session",
        "description": "Revoquer un jeton deja absent n'est pas une erreur.",
        "security": [
          {
            "SessionAgent": []
          }
        ],
        "responses": {
          "204": {
            "description": "Session revoquee"
          }
        }
      }
    },
    "/v1/auth/moi": {
      "get": {
        "summary": "Titulaire de la session",
        "description": "Permet au frontend de savoir au demarrage si son jeton vaut encore.",
        "security": [
          {
            "SessionAgent": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succes"
          }
        }
      }
    },
    "/v1/dossiers": {
      "get": {
        "summary": "Lister les dossiers de l'agence",
        "description": "Un agent voit les dossiers de son agence et n'en voit aucun autre.",
        "security": [
          {
            "SessionAgent": []
          }
        ],
        "parameters": [
          {
            "name": "statut",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "brouillon",
                "en_instruction",
                "accorde",
                "refuse",
                "annule"
              ]
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes"
          }
        }
      },
      "post": {
        "summary": "Creer un dossier",
        "description": "Le dossier nait en brouillon. La reference est unique par agence.",
        "security": [
          {
            "SessionAgent": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dossier"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dossier cree"
          },
          "409": {
            "description": "Reference deja employee"
          }
        }
      }
    },
    "/v1/dossiers/{id}": {
      "get": {
        "summary": "Lire un dossier et son historique",
        "description": "",
        "security": [
          {
            "SessionAgent": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes"
          }
        }
      },
      "patch": {
        "summary": "Modifier un dossier",
        "description": "Seul un brouillon est modifiable : une fois transmis a l'instruction, la decision doit porter sur ce qui a ete instruit.",
        "security": [
          {
            "SessionAgent": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Dossier"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dossier modifie"
          },
          "409": {
            "description": "Dossier fige ou reference prise"
          }
        }
      }
    },
    "/v1/dossiers/{id}/statut": {
      "post": {
        "summary": "Changer le statut",
        "description": "Le cycle de vie est une machine a etats : brouillon vers en_instruction ou annule ; en_instruction vers accorde, refuse ou annule ; accorde, refuse et annule sont terminaux.",
        "security": [
          {
            "SessionAgent": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "statut"
                ],
                "properties": {
                  "statut": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 1000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Statut change"
          },
          "409": {
            "description": "Transition refusee"
          }
        }
      }
    },
    "/v1/dossiers/{id}/echeancier": {
      "get": {
        "summary": "Echeancier du dossier",
        "description": "Calcule l'echeancier a partir des parametres du dossier et y rattache la simulation produite, qui porte le verdict de taux excessif.",
        "security": [
          {
            "SessionAgent": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Succes"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cleAPI": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Obligatoire en production ; le serveur refuse de demarrer sans elle."
      },
      "SessionAgent": {
        "type": "http",
        "scheme": "bearer",
        "description": "Jeton de session obtenu par POST /v1/auth/connexion. Le frontend etant servi depuis une autre origine, la session voyage dans l'en-tete Authorization plutot que dans un cookie : un jeton porteur n'est jamais envoye par le navigateur de lui-meme, ce qui supprime le risque CSRF. Le garder en memoire, non dans localStorage."
      }
    },
    "schemas": {
...
      "Bareme": {
        "type": "object",
        "description": "Taux effectif moyen applique, tel que publie par arrete.",
        "required": [
          "categorie",
          "semestre",
          "tem",
          "arrete",
          "publie_le"
        ],
        "properties": {
          "categorie": {
            "type": "string",
            "example": "credits_consommation"
          },
          "semestre": {
            "type": "string",
            "description": "Semestre d'application, forme 2026S1.",
            "example": "2026S1"
          },
          "tem": {
            "type": "string",
            "description": "Rendu en chaine : la valeur ne passe jamais par un flottant.",
            "example": "11.23"
          },
          "arrete": {
            "type": "string",
            "example": "Arrete du 28 juillet 2026"
          },
          "publie_le": {
            "type": "string",
            "format": "date",
            "example": "2026-07-28"
          }
        }
      },
...
    }
  }
}